Added explicit double cast for value/weight

+++ src/Algorithms/Collections/KnapsackItem.cs
@@ -10,7 +10,7 @@
         {
             Value = value;
             Weight = weight;
-            ValuePerWeight = value / weight;
+            ValuePerWeight = (double)value / weight;
         }
     }
 }
